app/requirements.txt
SQLAlchemy~=2.0.36
pydantic~=2.9.2
fastapi~=0.115.4
PyJWT~=2.9.0
passlib~=1.7.4
requests~=2.32.3
uvicorn~=0.32.0